gesture_mouse_control/requirements.txt
opencv-python>=4.5.0,<5.0.0
mediapipe==0.10.21
pyautogui>=0.9.53
pyyaml>=6.0
win10toast>=0.9
numpy>=1.19.0,<2.0.0
# Python 3.12 or lower is required due to mediapipe compatibility